requirements.txt
absl-py>=0.12.0
numpy>=1.19.2
scipy>=1.7.1
cvxopt>=1.2.6
dataclasses==0.8;python_version=="3.6"
dm-control>=0.0.364896371
dm-env==1.5
dm-haiku==0.0.4
dm-launchpad-nightly==0.3.0.dev20220413
dm-reverb>=0.4.0
dm-sonnet==2.0.0
dm-tree==0.1.6
jax==0.2.17;python_version=="3.6"
jaxlib==0.1.68;python_version=="3.6"
jax>=0.3.14;python_version>"3.6"
jaxlib>=0.3.14;python_version>"3.6"
keras>=2.6.0
optax==0.0.9
Pillow>=9.0.0
pytype==2022.6.30
pytest-xdist==2.3.0
rlax>=0.0.4
tensorflow-datasets>=4.4.0
tensorflow-estimator>=2.6.0
tensorflow>=2.6.0
tfp-nightly==0.14.0.dev20210818
trfl>=1.2.0
dm-acme>=0.2.2
dm-acme[jax]>=0.2.2
dm-acme[tensorflow]>=0.2.2
dm-acme[launchpad]>=0.2.2
dm-acme[envs]>=0.2.2
matplotlib==3.4.3
pandas==2.2.1
gym==0.26.2
